backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
gitpython>=3.1.40
databricks-sdk>=0.20.0
python-multipart>=0.0.6
